`constrain` renamed to `constraint`; cleanup
