+++ src/Mapper/TagMapper.php
@@ -5,9 +5,6 @@ namespace App\Mapper;
 use Cycle\Annotated\Annotation\Column;
 use Cycle\Annotated\Annotation\Table;
 use Cycle\ORM\Command\ContextCarrierInterface;
-use Cycle\ORM\Command\CommandInterface;
-use Cycle\ORM\Command\Database\Update;
-use Cycle\ORM\Context\ConsumerInterface;
 use Cycle\ORM\Heap\Node;
 use Cycle\ORM\Heap\State;
 use Cycle\ORM\Mapper\Mapper;
